fix(settings): fixed a bug causing the application to crash
