--- app/src/main/java/com/example/playlistmaker/domain/settings/model/IntentType.kt
@@ -1,11 +1,0 @@
-package com.example.playlistmaker.domain.settings.model
-
-sealed class IntentType {
-
-    data object ShareAppType : IntentType()
-
-    data object WriteToSupportType : IntentType()
-
-    data object ReadUserAgreementType : IntentType()
-
-}
